+++ src/connection/connection.rs
@@ -70,6 +70,7 @@ use crate::FourTupleIter;
 use crate::MultipathConfig;
 use crate::PacketInfo;
 use crate::PathEvent;
+use crate::PathStats;
 use crate::RecoveryConfig;
 use crate::Result;
 use crate::Shutdown;
@@ -1513,12 +1514,14 @@ impl Connection {
 
         // Prepare and encode packet header (except for the Length and Packet Number field)
         let space_id = self.get_space_id(pkt_type, path_id)?;
-        let pkt_num = self
-            .spaces
-            .get_mut(space_id)
-            .ok_or(Error::InternalError)?
-            .next_pkt_num;
-        let pkt_num_len = packet::packet_num_len(pkt_num)?;
+        let (pkt_num, pkt_num_len) = {
+            let space = self.spaces.get_mut(space_id).ok_or(Error::InternalError)?;
+            let largest_acked = space.get_largest_acked_pkt();
+            let pkt_num = space.next_pkt_num;
+            let pkt_num_len = packet::packet_num_len(pkt_num, largest_acked);
+            (pkt_num, pkt_num_len)
+        };
+
         let dcid_seq = self
             .paths
             .get(path_id)?
@@ -1573,7 +1576,7 @@ impl Connection {
         }
 
         // Encode packet number
-        let len = packet::encode_packet_num(pkt_num, &mut out[pkt_num_offset..left])?;
+        let len = packet::encode_packet_num(pkt_num, pkt_num_len, &mut out[pkt_num_offset..left])?;
         let payload_offset = pkt_num_offset + len;
 
         // Write frames into the packet payload
@@ -3582,6 +3585,19 @@ impl Connection {
         self.paths.get_active()
     }
 
+    /// Return an mutable reference to the specified path
+    pub fn get_path_stats(
+        &mut self,
+        local_addr: SocketAddr,
+        remote_addr: SocketAddr,
+    ) -> Result<&crate::PathStats> {
+        let pid = self
+            .paths
+            .get_path_id(&(local_addr, remote_addr))
+            .ok_or(Error::InvalidOperation("not found".into()))?;
+        Ok(self.paths.get_mut(pid)?.stats())
+    }
+
     /// Migrates the connection to the specified path.
     #[doc(hidden)]
     pub fn migrate_path(&mut self, local_addr: SocketAddr, remote_addr: SocketAddr) -> Result<()> {
@@ -4266,8 +4282,8 @@ pub(crate) mod tests {
             server_config: &mut Config,
             server_name: &str,
         ) -> Result<TestPair> {
-            let mut cli_cid_gen = RandomConnectionIdGenerator::new(client_config.cid_len, None);
-            let mut srv_cid_gen = RandomConnectionIdGenerator::new(server_config.cid_len, None);
+            let mut cli_cid_gen = RandomConnectionIdGenerator::new(client_config.cid_len);
+            let mut srv_cid_gen = RandomConnectionIdGenerator::new(server_config.cid_len);
             let client_scid = cli_cid_gen.generate();
             let server_scid = srv_cid_gen.generate();
             let client_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 9443);